Create a license report aggregator
This project generates a license report for Micronaut projects.
To do this, it will checkout the different projects to generate
a report for, and inject into the builds a license report plugin.

For this it makes use of the Gradle Tooling API, in order to
maximize compatibility of the different projects.

The reports can be generated by calling:

`./gradlew allReports`

and a ZIP with the reports can be generated using:

`./gradlew licenseReportZip`